replaced float with int in the callers

+++ Assets/Scripts/UI/UIManager.cs
@@ -118,7 +118,7 @@ public class UIManager : MonoBehaviour
 
         if (scoreManagerSO != null)
         {
-            float score = ScoreManagerSO.CalculateOverallTeamScore();
+            int score = ScoreManagerSO.CalculateOverallTeamScore();
             scoreText.text = score.ToString();
         }
